crowbar-backup: Remove cleanup and purge commands
These commands should really not belong to the crowbar-backup script:
blindly removing files and packages is not something we should do in a
product.

Somebody in the real world who will want to start fresh would install a
new OS and restore from backup, not purge like this.

Since it's mostly needed by CI, this has been moved there instead.