Increase resrore timeout
With the switch to the merged crowbar repositories, installing
crowbar takes a bit longer than it did before. Until this is fixed, we
need to increase the timeout to make sure the restore process does not
fail here. Set the timeout back to 300 again after the original issue
is fixed. See also https://bugzilla.suse.com/show_bug.cgi?id=942849